7_document_classifier/backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
python-multipart
pdfplumber
reportlab
pytest
docx2txt
python-docx